feat: add some new features

+++ Adjust/Source/Adjust/Public/AdjustAdRevenue.h
@@ -16,39 +16,30 @@ struct FAdjustAdRevenue
     GENERATED_USTRUCT_BODY()
 
 public:
-    // tested [ios, android]
     UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Adjust")
     FString Source;
 
-    // tested [ios, android]
     UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Adjust")
-    float Revenue = 0.0f;
+    double Revenue = 0.0;
 
-    // tested [ios, android]
     UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Adjust")
     FString Currency;
 
-    // tested [ios, android]
     UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Adjust")
     int AdImpressionsCount = 0;
 
-    // tested [ios, android]
     UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Adjust")
     FString AdRevenueNetwork;
 
-    // tested [ios, android]
     UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Adjust")
     FString AdRevenueUnit;
 
-    // tested [ios, android]
     UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Adjust")
     FString AdRevenuePlacement;
 
-    // tested [ios, android]
     UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Adjust")
     TMap<FString, FString> CallbackParameters;
 
-    // tested [ios, android]
     UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Adjust")
     TMap<FString, FString> PartnerParameters;
 };
